Status Change - Denied
This is a project for development only. Development is not something that the BCF funds as this does not directly increase adoption and awareness of Bitcoin Cash. The cost of development also puts this outside the budget of what the BCF can afford.

Unfortunately this means we are unable to fund this project.